Generic invocation supports generalizer.Generalizer
