+++ generic/default/go-client/conf/client.yml
