+++ generic/default/go-client/conf/log.yml
