+++ generic/default/go-client/pkg/user.go
