+++ generic/default/go-server/cmd/server.go
@@ -18,7 +18,6 @@
 package main
 
 import (
-	"context"
 	"fmt"
 	"os"
 	"os/signal"
@@ -27,53 +26,39 @@ import (
 )
 
 import (
-	"github.com/apache/dubbo-go-samples/generic/go-client/pkg"
-	"github.com/dubbogo/gost/log"
+	hessian "github.com/apache/dubbo-go-hessian2"
+	pkg2 "github.com/apache/dubbo-go-samples/generic/default/go-server/pkg"
 )
 
 import (
+
 	_ "dubbo.apache.org/dubbo-go/v3/cluster/cluster_impl"
 	_ "dubbo.apache.org/dubbo-go/v3/cluster/loadbalance"
 	"dubbo.apache.org/dubbo-go/v3/common/logger"
 	_ "dubbo.apache.org/dubbo-go/v3/common/proxy/proxy_factory"
 	"dubbo.apache.org/dubbo-go/v3/config"
 	_ "dubbo.apache.org/dubbo-go/v3/filter/filter_impl"
-	"dubbo.apache.org/dubbo-go/v3/protocol/dubbo"
+	_ "dubbo.apache.org/dubbo-go/v3/protocol/dubbo"
 	_ "dubbo.apache.org/dubbo-go/v3/registry/protocol"
 	_ "dubbo.apache.org/dubbo-go/v3/registry/zookeeper"
 )
 
 var (
-	appName         = "UserConsumer"
-	referenceConfig = config.ReferenceConfig{
-		InterfaceName: "org.apache.dubbo.UserProvider",
-		Cluster:       "failover",
-		Registry:      "demoZk",
-		Protocol:      dubbo.DUBBO,
-		Generic:       true,
-	}
+	survivalTimeout = int(3e9)
 )
 
-func init() {
-	config.Load()
-	referenceConfig.GenericLoad(appName) //appName is the unique identification of RPCService
-	time.Sleep(3 * time.Second)
-}
-
-// need to setup environment variable "CONF_CONSUMER_FILE_PATH" to "conf/client.yml" before run
+// need to setup environment variable "CONF_PROVIDER_FILE_PATH" to "conf/server.yml" before run
 func main() {
-	gxlog.CInfo("\n\ncall getUser")
-	callGetUser()
-	gxlog.CInfo("\n\ncall queryUser")
-	callQueryUser()
+	hessian.RegisterPOJO(&pkg2.User{})
+	config.Load()
+
 	initSignal()
 }
 
 func initSignal() {
 	signals := make(chan os.Signal, 1)
 	// It is not possible to block SIGKILL or syscall.SIGSTOP
-	signal.Notify(signals, os.Interrupt, os.Kill, syscall.SIGHUP,
-		syscall.SIGQUIT, syscall.SIGTERM, syscall.SIGINT)
+	signal.Notify(signals, os.Interrupt, os.Kill, syscall.SIGHUP, syscall.SIGQUIT, syscall.SIGTERM, syscall.SIGINT)
 	for {
 		sig := <-signals
 		logger.Infof("get signal %s", sig.String())
@@ -81,55 +66,14 @@ func initSignal() {
 		case syscall.SIGHUP:
 			// reload()
 		default:
-			time.AfterFunc(10*time.Second, func() {
+			time.AfterFunc(time.Duration(survivalTimeout), func() {
 				logger.Warnf("app exit now by force...")
 				os.Exit(1)
 			})
 
 			// The program exits normally or timeout forcibly exits.
-			fmt.Println("app exit now...")
+			fmt.Println("provider app exit now...")
 			return
 		}
 	}
-}
-
-func callGetUser() {
-	gxlog.CInfo("\n\n\nstart to generic invoke")
-	resp, err := referenceConfig.GetRPCService().(*config.GenericService).Invoke(
-		context.TODO(),
-		[]interface{}{
-			"GetUser",
-			[]string{"java.lang.String"},
-			[]interface{}{"A003"},
-		},
-	)
-	if err != nil {
-		panic(err)
-	}
-	gxlog.CInfo("res: %+v\n", resp)
-	gxlog.CInfo("success!")
-
-}
-func callQueryUser() {
-	gxlog.CInfo("\n\n\nstart to generic invoke")
-	user := pkg.User{
-		ID:   "3213",
-		Name: "panty",
-		Age:  25,
-		Time: time.Now(),
-	}
-	resp, err := referenceConfig.GetRPCService().(*config.GenericService).Invoke(
-		context.TODO(),
-		[]interface{}{
-			"queryUser",
-			[]string{"org.apache.dubbo.User"},
-			[]interface{}{user},
-		},
-	)
-	if err != nil {
-		panic(err)
-	}
-	gxlog.CInfo("res: %+v\n", resp)
-	gxlog.CInfo("success!")
-
 }
