+++ generic/default/go-server/conf/client.yml
