+++ generic/default/go-server/conf/server.yml
