+++ generic/default/go-server/docker/docker-compose.yml
