+++ generic/default/go-server/tests/integration/main_test.go
@@ -43,7 +43,7 @@ var referenceConfig = config.ReferenceConfig{
 	Cluster:       "failover",
 	Registry:      "demoZk",
 	Protocol:      dubbo.DUBBO,
-	Generic:       true,
+	Generic:       "true",
 }
 
 func TestMain(m *testing.M) {
