+++ generic/default/java-client/2.6/src/main/java/org/apache/dubbo/User.java
