+++ generic/default/java-client/2.6/src/main/resources/log4j.properties
