+++ generic/default/java-client/2.7/pom.xml
