+++ generic/default/java-client/2.7/src/main/resources/log4j.properties
