+++ generic/default/java-server/2.6/src/main/resources/META-INF/spring/dubbo.provider.xml
