+++ generic/default/java-server/2.7/src/main/resources/META-INF/spring/dubbo.provider.xml
